forgery_detector/requirements.txt
streamlit
pillow
opencv-python
numpy
torch
torchvision
exifread
pymupdf
matplotlib
scikit-learn